+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @baunilha/tokens
 
+## 1.0.2
+
+### Patch Changes
+
+- Fixed npm files
+
 ## 1.0.1
 
 ### Patch Changes